+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - [x] made the interactive image viewer blink less ~ to none -- making it easier to the eye
 - [x] added a `--no-linenumber` flag to remove line numbers from the markdown viewer
 - [x] raw text from stdin now defaults to markdown instead of txt in the markdown viewer
-- [ ] improved rendering of images in tmux by moving the cursor after the image
+- [x] improved rendering of images in tmux by moving the cursor after the image
 - [ ] added an optional config file, allowing users to have a config file instead of aliasing 
 - [x] improved GP support auto detetion -- especially in tmux
 - [ ] the ls command now combines images by row to fix bugs from quick image printing